Add role members validation to ensure a correct format used
An IAM policy objects in Google Cloud consist of a list of role bindings,
and a particular role binding expects a list of principals (members) to
be provided.

When provided, the list of principals is to be specified using a special
format, where each principal is to be prefixed with a specific prefix
such as e.g., "user:", "group:", etc., depending on given principal
type.  Should this prefix be absent, then the Google API will return an
error resulting in Terraform run failure.

Thus, add validation of the "members" variable to ensure that each
principal given has a valid prefix set.

Signed-off-by: Krzysztof Wilczyński <kw@linux.com>